+++ MoreShipUpgrades/Misc/LategameConfiguration.cs
@@ -35,6 +35,7 @@ namespace MoreShipUpgrades.Misc
     public class LategameConfiguration : SyncedConfig2<LategameConfiguration>
     {
         #region Enabled
+        [field: SyncedEntryField] public SyncedEntry<bool> SCRAP_KEEPER_ENABLED {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<bool> PARTICLE_INFUSER_ENABLED {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<bool> SILVER_BULLETS_ENABLED {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<bool> FUSION_MATTER_ENABLED {  get; set; }
@@ -126,6 +127,7 @@ namespace MoreShipUpgrades.Misc
         #endregion
 
         #region Initial Prices
+        [field: SyncedEntryField] public SyncedEntry<int> SCRAP_KEEPER_PRICE {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<int> PARTICLE_INFUSER_PRICE {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<int> SILVER_BULLETS_PRICE {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<int> FUSION_MATTER_PRICE {  get; set; }
@@ -185,6 +187,10 @@ namespace MoreShipUpgrades.Misc
         #endregion
 
         #region Attributes
+        [field: SyncedEntryField] public SyncedEntry<string> SCRAP_KEEPER_PRICES {  get; set; }
+        [field: SyncedEntryField] public SyncedEntry<string> SCRAP_KEEPER_OVERRIDE_NAME {  get; set; }
+        [field: SyncedEntryField] public SyncedEntry<int> SCRAP_KEEPER_INITIAL_KEEP_SCRAP_CHANCE_INCREASE {  get; set; }
+        [field: SyncedEntryField] public SyncedEntry<int> SCRAP_KEEPER_INCREMENTAL_KEEP_SCRAP_CHANCE_INCREASE { get; set; }
         [field: SyncedEntryField] public SyncedEntry<string> PARTICLE_INFUSER_PRICES {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<string> PARTICLE_INFUSER_OVERRIDE_NAME {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<int> PARTICLE_INFUSER_INITIAL_TELEPORT_SPEED_INCREASE { get; set; }
@@ -514,6 +520,7 @@ namespace MoreShipUpgrades.Misc
         [field: SyncedEntryField] public SyncedEntry<string> ITEM_PROGRESSION_APPARATICE_ITEMS {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<bool> ITEM_PROGRESSION_NO_PURCHASE_UPGRADES {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<bool> ITEM_PROGRESSION_ALWAYS_SHOW_ITEMS {  get; set; }
+        [field: SyncedEntryField] public SyncedEntry<string> SCRAP_KEEPER_ITEM_PROGRESSION_ITEMS {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<string> PARTICLE_INFUSER_ITEM_PROGRESSION_ITEMS {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<string> SILVER_BULLETS_ITEM_PROGRESSION_ITEMS {  get; set; }
         [field: SyncedEntryField] public SyncedEntry<string> FUSION_MATTER_ITEM_PROGRESSION_ITEMS {  get; set; }
@@ -623,6 +630,7 @@ namespace MoreShipUpgrades.Misc
 
             topSection = LguConstants.OVERRIDE_NAMES_SECTION;
             OVERRIDE_UPGRADE_NAMES              = cfg.BindSyncedEntry(topSection, LguConstants.OVERRIDE_NAMES_ENABLED_KEY, LguConstants.OVERRIDE_NAMES_ENABLED_DEFAULT, LguConstants.OVERRIDE_NAMES_ENABLED_DESCRIPTION);
+            SCRAP_KEEPER_OVERRIDE_NAME          = cfg.BindSyncedEntry(topSection, LguConstants.SCRAP_KEEPER_OVERRIDE_NAME_KEY, ScrapKeeper.UPGRADE_NAME);
             PARTICLE_INFUSER_OVERRIDE_NAME      = cfg.BindSyncedEntry(topSection, LguConstants.PARTICLE_INFUSER_OVERRIDE_NAME_KEY, ParticleInfuser.UPGRADE_NAME);
             SILVER_BULLETS_OVERRIDE_NAME        = cfg.BindSyncedEntry(topSection, LguConstants.SILVER_BULLETS_OVERRIDE_NAME_KEY, SilverBullets.UPGRADE_NAME);
             FUSION_MATTER_OVERRIDE_NAME         = cfg.BindSyncedEntry(topSection, LguConstants.FUSION_MATTER_OVERRIDE_NAME_KEY, FusionMatter.UPGRADE_NAME);
@@ -725,6 +733,18 @@ namespace MoreShipUpgrades.Misc
             #endregion
 
             #region Upgrades
+
+            #region Scrap Keeper
+
+            topSection = ScrapKeeper.UPGRADE_NAME;
+            SCRAP_KEEPER_ENABLED = cfg.BindSyncedEntry(topSection, LguConstants.SCRAP_KEEPER_ENABLED_KEY, LguConstants.SCRAP_KEEPER_ENABLED_DEFAULT, LguConstants.SCRAP_KEEPER_ENABLED_DESCRIPTION);
+            SCRAP_KEEPER_PRICE = cfg.BindSyncedEntry(topSection, LguConstants.SCRAP_KEEPER_PRICE_KEY, LguConstants.SCRAP_KEEPER_PRICE_DEFAULT);
+            SCRAP_KEEPER_PRICES = cfg.BindSyncedEntry(topSection, BaseUpgrade.PRICES_SECTION, ScrapKeeper.PRICES_DEFAULT, BaseUpgrade.PRICES_DESCRIPTION);
+            SCRAP_KEEPER_INITIAL_KEEP_SCRAP_CHANCE_INCREASE = cfg.BindSyncedEntry(topSection, LguConstants.SCRAP_KEEPER_INITIAL_KEEP_SCRAP_CHANCE_INCREASE_KEY, LguConstants.SCRAP_KEEPER_INITIAL_KEEP_SCRAP_CHANCE_INCREASE_DEFAULT, LguConstants.SCRAP_KEEPER_INITIAL_KEEP_SCRAP_CHANCE_INCREASE_DESCRIPTION);
+            SCRAP_KEEPER_INCREMENTAL_KEEP_SCRAP_CHANCE_INCREASE = cfg.BindSyncedEntry(topSection, LguConstants.SCRAP_KEEPER_INCREMENTAL_KEEP_SCRAP_CHANCE_INCREASE_KEY, LguConstants.SCRAP_KEEPER_INCREMENTAL_KEEP_SCRAP_CHANCE_INCREASE_DEFAULT, LguConstants.SCRAP_KEEPER_INCREMENTAL_KEEP_SCRAP_CHANCE_INCREASE_DESCRIPTION);
+            SCRAP_KEEPER_ITEM_PROGRESSION_ITEMS = cfg.BindSyncedEntry(topSection, LguConstants.ITEM_PROGRESSION_ITEMS_KEY, LguConstants.ITEM_PROGRESSION_ITEMS_DEFAULT, LguConstants.ITEM_PROGRESSION_ITEMS_DESCRIPTION);
+
+            #endregion
 
             #region Particle Infuser