+++ ChangeLog.md
@@ -8,6 +8,7 @@
 - Implemented tier upgrade which increases the shotgun's damage output.
 - Implemented tier upgrade which increases the shotgun's overall range and its effective damage ranges.
 - Implemented tier upgrade which increases the teleporter's speed.
+- Implemented tier upgrade which in case of a full team wipe, each scrap item present in the ship has a chance of not being discarded.
 - Fixed issue with medkit on Extraction contract due to different names.
 
 </details>